Added PutObject with test support for encryption file
